+++ app/means_test/forms/about_you.py
@@ -3,7 +3,7 @@ from govuk_frontend_wtf.wtforms_widgets import GovTextInput
 from wtforms.validators import InputRequired, NumberRange
 from app.means_test.validators import ValidateIf
 from app.means_test.widgets import MeansTestRadioInput
-from flask_babel import gettext as _
+from flask_babel import lazy_gettext as _
 from app.means_test import YES, NO
 from app.means_test.forms import BaseMeansTestForm
