+++ app/static/src/js/scripts.js
@@ -1,2 +1,3 @@
 import './cookie-banner.js';
 import './custom.js';
+import './conditional-questions.js'
